Proccess 0 does the printing
I beleive that what I made is right, but it still doesn't pass the
longer tests. At least now the output is allways the same (before,
at least when I tried, it wasn't). I tried to find what is wrong,
but I didn't find the error. Please check what I've made and see
if it's ok.